backend/tasks.py
"""
Celery 任务定义
"""
import time
import random
from celery import Celery
from PIL import Image, ImageEnhance, ImageFilter
import io
import os

# 创建 Celery 实例
celery_app = Celery(
    'tasks',
    broker=os.getenv('REDIS_URL', 'redis://localhost:6379'),
    backend=os.getenv('REDIS_URL', 'redis://localhost:6379')
)

# 任务状态存储
task_progress = {}

@celery_app.task(bind=True)
def process_image(self, filename: str, operations: list):
    """
    处理图片的 Celery 任务
    
    Args:
        filename: 图片文件名
        operations: 操作列表 [{type: 'face_slim', intensity: 0.5}, ...]
    """
    task_id = self.request.id
    
    # 更新进度
    def update_progress(progress: int, status: str = "processing"):
        task_progress[task_id] = {
            "progress": progress,
            "status": status
        }
        self.update_state(
            state='PROGRESS',
            meta={'progress': progress, 'status': status}
        )
    
    try:
        # 加载图片
        update_progress(10, "正在加载图片")
        image_path = f"../uploads/{filename}"
        image = Image.open(image_path)
        
        # 模拟处理过程
        update_progress(25, "正在分析图片")
        time.sleep(1)  # 模拟分析时间
        
        # 应用各种效果
        for i, operation in enumerate(operations):
            progress = 25 + (50 / len(operations)) * (i + 1)
            
            if operation['type'] == 'face_slim':
                update_progress(int(progress), "正在进行瘦脸处理")
                # 模拟瘦脸效果（实际应调用 AI 模型）
                image = simulate_face_slim(image, operation['intensity'])
                
            elif operation['type'] == 'body_slim':
                update_progress(int(progress), "正在进行瘦身处理")
                # 模拟瘦身效果
                image = simulate_body_slim(image, operation['intensity'])
                
            elif operation['type'] == 'beauty_filter':
                update_progress(int(progress), "正在应用美颜滤镜")
                # 模拟美颜效果
                image = simulate_beauty_filter(
                    image, 
                    operation['filter_type'], 
                    operation['intensity']
                )
            
            time.sleep(1)  # 模拟处理时间
        
        # 保存结果
        update_progress(90, "正在保存结果")
        result_filename = f"result_{task_id}.jpg"
        result_path = f"../results/{result_filename}"
        image.save(result_path, "JPEG", quality=95)
        
        update_progress(100, "completed")
        
        return {
            "status": "completed",
            "result_filename": result_filename,
            "result_url": f"/api/result/{task_id}"
        }
        
    except Exception as e:
        update_progress(0, "failed")
        return {
            "status": "failed",
            "error": str(e)
        }

def simulate_face_slim(image: Image.Image, intensity: float) -> Image.Image:
    """模拟瘦脸效果（临时实现）"""
    # 这里只是简单地稍微压缩图片宽度来模拟瘦脸
    width, height = image.size
    new_width = int(width * (1 - intensity * 0.05))  # 最多压缩5%
    return image.resize((new_width, height), Image.Resampling.LANCZOS)

def simulate_body_slim(image: Image.Image, intensity: float) -> Image.Image:
    """模拟瘦身效果（临时实现）"""
    # 简单的整体压缩
    width, height = image.size
    new_width = int(width * (1 - intensity * 0.08))  # 最多压缩8%
    return image.resize((new_width, height), Image.Resampling.LANCZOS)

def simulate_beauty_filter(image: Image.Image, filter_type: str, intensity: float) -> Image.Image:
    """模拟美颜滤镜（临时实现）"""
    # 应用不同的滤镜效果
    if filter_type == 'natural':
        # 自然美颜：轻微模糊 + 提亮
        image = image.filter(ImageFilter.SMOOTH_MORE)
        enhancer = ImageEnhance.Brightness(image)
        image = enhancer.enhance(1 + intensity * 0.2)
        
    elif filter_type == 'glamour':
        # 魅力滤镜：增加对比度和饱和度
        enhancer = ImageEnhance.Contrast(image)
        image = enhancer.enhance(1 + intensity * 0.3)
        enhancer = ImageEnhance.Color(image)
        image = enhancer.enhance(1 + intensity * 0.2)
        
    elif filter_type == 'fresh':
        # 清新滤镜：提亮 + 降低饱和度
        enhancer = ImageEnhance.Brightness(image)
        image = enhancer.enhance(1 + intensity * 0.3)
        enhancer = ImageEnhance.Color(image)
        image = enhancer.enhance(1 - intensity * 0.2)
        
    elif filter_type == 'artistic':
        # 艺术滤镜：边缘增强
        image = image.filter(ImageFilter.EDGE_ENHANCE)
        
    return image